+++ packages/core/navigation/examples/utils/react-router/PageNavigation.js
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import AtlassianIcon from '@atlaskit/logo';
+import { AtlassianIcon } from '@atlaskit/logo';
 import Lorem from 'react-lorem-component';
 import Page from '@atlaskit/page';
 import Navigation, { AkContainerTitle } from '../../../src';