+++ packages/core/page/examples/04-navigation-and-banner-integration.js
@@ -3,7 +3,7 @@ import React, { Component, Fragment } from 'react';
 import Navigation, { AkNavigationItem } from '@atlaskit/navigation';
 import Banner from '@atlaskit/banner';
 import Button from '@atlaskit/button';
-import AtlassianIcon from '@atlaskit/logo';
+import { AtlassianIcon } from '@atlaskit/logo';
 import Page, { Grid } from '../src';
 
 type State = {