+++ packages/core/select/examples/18-element-before.js
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import AtlassianIcon from '@atlaskit/logo';
+import { AtlassianIcon } from '@atlaskit/logo';
 import Select from '../src';
 
 const formatOptionLabel = (option, { context }) => {